Extract sample size computer
