+++ spotify_confidence/analysis/frequentist/sample_size_calculator.py
@@ -2,8 +2,7 @@ from typing import Union, Iterable, Tuple
 
 from pandas import DataFrame
 
-from spotify_confidence.analysis.frequentist.confidence_computers.generic_computer import GenericComputer
-from ..abstract_base_classes.confidence_computer_abc import ConfidenceComputerABC
+from spotify_confidence.analysis.frequentist.confidence_computers.sample_size_computer import SampleSizeComputer
 from ..confidence_utils import (
     listify,
 )
@@ -20,34 +19,21 @@ class SampleSizeCalculator:
         categorical_group_columns: Union[None, str, Iterable] = None,
         interval_size: float = 0.95,
         correction_method: str = BONFERRONI,
-        confidence_computer: ConfidenceComputerABC = None,
         metric_column=None,
         power: float = 0.8,
     ):
-        if confidence_computer is not None:
-            self._confidence_computer = confidence_computer
-        else:
-            self._confidence_computer = GenericComputer(
-                data_frame=data_frame.assign(**{METHOD_COLUMN_NAME: ZTEST}),
-                numerator_column=None,
-                numerator_sum_squares_column=None,
-                denominator_column=None,
-                categorical_group_columns=listify(categorical_group_columns),
-                ordinal_group_column=None,
-                interval_size=interval_size,
-                correction_method=correction_method.lower(),
-                method_column=METHOD_COLUMN_NAME,
-                bootstrap_samples_column=None,
-                metric_column=metric_column,
-                treatment_column=None,
-                power=power,
-                point_estimate_column=point_estimate_column,
-                var_column=var_column,
-                is_binary_column=is_binary_column,
-                feature_column=None,
-                feature_sum_squares_column=None,
-                feature_cross_sum_column=None,
-            )
+        self._sample_size_computer = SampleSizeComputer(
+            data_frame=data_frame.assign(**{METHOD_COLUMN_NAME: ZTEST}),
+            categorical_group_columns=listify(categorical_group_columns),
+            interval_size=interval_size,
+            correction_method=correction_method.lower(),
+            method_column=METHOD_COLUMN_NAME,
+            metric_column=metric_column,
+            power=power,
+            point_estimate_column=point_estimate_column,
+            var_column=var_column,
+            is_binary_column=is_binary_column,
+        )
 
     def sample_size(
         self,
@@ -76,7 +62,7 @@ class SampleSizeCalculator:
             difference of mde or and absolute difference of mde*avg. If nim is set the null hypothesis is -nim*avg
             if preferred_direction is "increase" and +nim*avg if preferred direction is "decrease"
         """
-        return self._confidence_computer.compute_sample_size(
+        return self._sample_size_computer.compute_sample_size(
             treatment_weights, mde_column, nim_column, preferred_direction_column, final_expected_sample_size_column
         )
 
@@ -90,7 +76,7 @@ class SampleSizeCalculator:
             Tuple (list, int), where the list contains optimal weights and the int is the sample size that would be
             required if those weights were used
         """
-        return self._confidence_computer.compute_optimal_weights_and_sample_size(sample_size_df, number_of_groups)
+        return self._sample_size_computer.compute_optimal_weights_and_sample_size(sample_size_df, number_of_groups)
 
     def powered_effect(
         self,
@@ -116,6 +102,6 @@ class SampleSizeCalculator:
         consistent with the sample size calculations, i.e. if you first calculate a sample size for a
         specific MDE and then calculate the powered effect for that sample size, the original MDE will be returned.
         """
-        return self._confidence_computer.compute_powered_effect(
+        return self._sample_size_computer.compute_powered_effect(
             treatment_weights, mde_column, nim_column, preferred_direction_column, sample_size
         )
